23680961: rename usecase and delivery
