+++ delivery/iapi.go
@@ -1,4 +1,4 @@
-package usecase
+package delivery
 
 import "net/http"
