+++ delivery/requests.go
@@ -1,4 +1,4 @@
-package usecase
+package delivery
 
 type SignupRequest struct {
 	Login     string `json:"login"`
