+++ usecase/core_test.go
@@ -1,4 +1,4 @@
-package delivery
+package usecase
 
 import (
 	"testing"
